requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
nltk==3.8.1
streamlit==1.32.2
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.4
python-dotenv==1.0.1